Add Svelte Wallet Adapter for Solana
Follow the steps at https://github.com/svelte-on-solana/wallet-adapter
to add a Solana Wallet Button to the app.

While installing dependencies on `@solana/web3.js`, I found that the
latest version introduces a transitive dependency on React (by way of
React Native). This dependency was introduced in v1.47.4, so I installed
the previous version of `@solana/web3.js`